+++ tests/testthat/test-funs_date-lazy.R
@@ -798,4 +798,24 @@ test_that("errors for durations", {
   expect_snapshot_lazy(mutate(test, x = minutes(1.2)), error = TRUE)
 })
 
+test_that("errors for rolling functions", {
+  test <- pl$LazyFrame(x = 1)
+  expect_snapshot_lazy(
+    mutate(test, x = rollbackward(x, roll_to_first = "a")),
+    error = TRUE
+  )
+  expect_snapshot_lazy(
+    mutate(test, x = rollbackward(x, preserve_hms = "a")),
+    error = TRUE
+  )
+  expect_snapshot_lazy(
+    mutate(test, x = rollforward(x, roll_to_first = "a")),
+    error = TRUE
+  )
+  expect_snapshot_lazy(
+    mutate(test, x = rollforward(x, preserve_hms = "a")),
+    error = TRUE
+  )
+})
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)